#414: factor out common parts from tests/lite.rs into tests/lite_tests
